+++ src/app/state.ts
@@ -5,7 +5,7 @@ import {nip19} from "nostr-tools"
 import {navigate} from "svelte-routing"
 import {writable} from "svelte/store"
 import {whereEq, omit, filter, pluck, sortBy, slice} from "ramda"
-import {hash, sleep, shuffle, doPipe} from "hurdak"
+import {hash, sleep, doPipe} from "hurdak"
 import {warn} from "src/util/logger"
 import {now} from "src/util/misc"
 import {userKinds, noteKinds} from "src/util/nostr"
@@ -229,8 +229,7 @@ export const publishWithToast = async (
 // Feeds
 
 export const compileFilter = (filter: DynamicFilter): Filter => {
-  const getAuthors = (pubkeys: string[]) =>
-    shuffle(pubkeys.length > 0 ? pubkeys : (Env.DEFAULT_FOLLOWS as string[])).slice(0, 256)
+  const getAuthors = (pubkeys: string[]) => (pubkeys.length > 0 ? pubkeys : Env.DEFAULT_FOLLOWS)
 
   if (filter.authors === "global") {
     filter = omit(["authors"], filter)